+++ app/youtubers/[channelId]/page.tsx
@@ -1,7 +1,7 @@
 import { compareDesc } from 'date-fns';
 import { notFound } from 'next/navigation';
-import { getAllChannels, getChannel } from '@/app/actions/fetchChannels';
 import type { YouTubeVideoItem } from '@/app/actions/fetchYouTube';
+import { getAllChannels, getChannel } from '@/app/actions/handleYTData';
 import Blog from '@/components/Blog/Blog';
 import { siteConfig } from '@/config/site';
 import { absoluteUrl } from '@/src/util/utils';
